# backend/static/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/generate": {
      "post": {
        "operationId": "AvatarController_generateAvatarV3",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateAvatarV3Dto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Emoji avatar generated successfully"
          }
        },
        "summary": "Generate emoji avatar (API v3)",
        "tags": ["Avatar"]
      }
    },
...
    "/list": {
      "get": {
        "operationId": "AvatarController_listAvatars",
        "parameters": [
          {
            "name": "pick",
            "required": false,
            "in": "query",
            "description": "Number of records to retrieve (default: 10, max: 100)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of records to skip (default: 0)",
            "schema": {
              "minimum": 0,
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Avatar list retrieved successfully"
          },
          "204": {
            "description": "No avatars found"
          }
        },
        "summary": "Get list of avatars with pagination",
        "tags": ["Avatar"]
      }
    },
    "/{id}": {
      "get": {
        "operationId": "AvatarController_getAvatar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Avatar ID (UUID)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filter to apply (grayscale, sepia, negative)",
            "schema": {
              "example": "grayscale",
              "type": "string",
              "enum": ["grayscale", "sepia", "negative"]
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "description": "Size parameter (4-9, where 2^n)",
            "schema": {
              "minimum": 4,
              "maximum": 9,
              "example": 6,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Avatar retrieved successfully"
          },
          "400": {
            "description": "Bad request - invalid parameters"
          },
          "404": {
            "description": "Avatar not found"
          }
        },
        "summary": "Get avatar by ID",
        "tags": ["Avatar"]
      },
      "delete": {
        "operationId": "AvatarController_deleteAvatar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Avatar ID (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Avatar deleted successfully"
          }
        },
        "summary": "Delete avatar by ID",
        "tags": ["Avatar"]
      }
    },
...
  },
  "info": {
    "title": "Avatar Generation API",
    "description": "API for generating and managing avatars similar to GitHub/GitLab",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "GenerateAvatarV3Dto": {
        "type": "object",
        "properties": {
          "emoji": {
            "type": "string",
            "description": "Unicode emoji character",
            "example": "😀",
            "maxLength": 20
          },
          "backgroundType": {
            "type": "string",
            "description": "Background type for the avatar",
            "enum": ["solid", "linear", "radial"],
            "example": "solid"
          },
          "primaryColor": {
            "type": "string",
            "description": "Primary color for background generation (hex format)",
            "example": "#3B82F6",
            "pattern": "^#[0-9A-Fa-f]{6}$"
          },
          "foreignColor": {
            "type": "string",
            "description": "Secondary color for gradient backgrounds (hex format)",
            "example": "#EF4444",
            "pattern": "^#[0-9A-Fa-f]{6}$"
          },
          "angle": {
            "type": "number",
            "description": "Gradient angle in degrees (0-360, only for linear background)",
            "minimum": 0,
            "maximum": 360,
            "example": 90
          },
          "emojiSize": {
            "type": "string",
            "description": "Size of the emoji relative to avatar",
            "enum": ["small", "medium", "large"],
            "example": "large",
            "default": "large"
          }
        },
        "required": ["emoji", "backgroundType"]
      }
    }
  }
}